+++ apps/react-example-nested-stores/src/app/store/effects/index.ts
@@ -1,5 +1,5 @@
-import { scan, startWith } from 'rxjs/operators';
-import { Effect } from '@rx-store/core';
+import { scan, startWith, ignoreElements } from 'rxjs/operators';
+import { RootEffect } from '@rx-store/core';
 import { RootContextValue } from '../../types';
 
 /**
@@ -19,9 +19,13 @@ import { RootContextValue } from '../../types';
  * by accessing the context value directly in your components, and
  * subscribing. See the <Provider /> component for an example.
  */
-export const appRootEffect: Effect<RootContextValue> = ({ sources, sinks }) =>
+export const appRootEffect: RootEffect<RootContextValue> = ({
+  sources,
+  sinks,
+}) =>
   sources.counterChange$().pipe(
     scan((acc, val) => acc + val, 0),
     startWith(0),
-    sinks.count$()
+    sinks.count$(),
+    ignoreElements()
   );